requirements.txt
# Library RAG System Dependencies
# For 老爷's Academic Paper Library

# Core RAG Framework
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-llms-anthropic>=0.1.0

# PDF Parsing - Nougat (Meta's Academic PDF Parser)
nougat-ocr>=0.1.17

# Vector Database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0
FlagEmbedding>=1.2.0  # BGE-M3 and reranker models

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
click>=8.0.0
pyyaml>=6.0.0
tenacity>=8.0.0

# Gemini API
google-genai>=1.0.0

# Optional: Streamlit UI
# streamlit>=1.30.0

# PyTorch - install separately with CUDA support:
# pip install torch --index-url https://download.pytorch.org/whl/cu121
# torch>=2.0.0
